requirements.txt
Flask==1.1.1
pyyaml